requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.15.0
asyncpg>=0.29.0

# Multi-Tenant Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Orchestration & Tracking
apache-airflow>=2.7.0
mlflow>=2.9.0
dvc>=3.30.0
dvc-gs>=3.0.0

# API & Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# Cloud Storage
google-cloud-storage>=2.13.0
google-cloud-aiplatform>=1.38.0

# Monitoring & Validation
prometheus-client>=0.19.0
evidently>=0.4.0
great-expectations>=0.18.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
requests>=2.31.0
